+++ Tests/ReferenceImages_64/StripePaymentSheetTests.VerticalPaymentMethodListViewControllerSnapshotTest/testAppearance@3x.png
