+++ Tests/ReferenceImages_64/StripePaymentSheetTests.VerticalPaymentMethodListViewControllerSnapshotTest/testDarkMode@3x.png
