+++ Tests/ReferenceImages_64/StripePaymentSheetTests.VerticalPaymentMethodListViewControllerSnapshotTest/testNoSavedPM_noApplePayLink@3x.png
