+++ Tests/ReferenceImages_64/StripePaymentSheetTests.VerticalPaymentMethodListViewControllerSnapshotTest/testSavedCard_ApplePayLink@3x.png
