+++ Tests/ReferenceImages_64/StripePaymentSheetTests.VerticalPaymentMethodListViewControllerSnapshotTest/testSavedCard_noApplePayLink@3x.png
